Implemented queue helper functions
Implemented _uik_queue_insert() and _uik_queue_next() for handling
task queues.

NOTE: this implementation will crash and burn if malloc() and free()
aren't present on the AVR. I'm a bit doubtful that free() actually
returns memory to the heap, or that there is any functionality for
compacting the heap. If this is the case, then taking tasks on and
off the ready queue will be a memory leak and this system will
end up clobbering it's system stack very quickly.

You have been warned.